requirements.txt
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
Deprecated==1.2.13
Flask==2.0.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
itsdangerous==2.1.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.1.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.2
python-decouple==3.6
six==1.16.0
Werkzeug==2.0.3
wrapt==1.13.3
